server1/backend/app/requirements.txt
asgiref==3.7.2
Django==5.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Markdown==3.5.2
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.9.0
django-environ==0.11.2
mysqlclient==2.2.1
gunicorn==21.2.0
django-cors-headers 
